+++ changelog.md
@@ -11,8 +11,8 @@
 * Deprecate `StatusError`, `raise` and `raiseStatus` (#351)
 * Add doctest, refactor some inline examples into doctests (#353)
 * document "`defaultHandler` only applies to endpoints defined after it" (#237)
-* add `setHeader1`, `addHeader1`, deprecate `setHeader`, `addHeader` (#94)
-* add `setCookie1`, `setSimpleCookie1` (#94)
+* add `setHeader1`, `addHeader1`, deprecate `setHeader`, `addHeader` (#92)
+* add `setCookie1`, `setSimpleCookie1` (#92)
 
 Breaking: